ttlens/requirements.txt
deprecation>=2.1.0
docopt>=0.6.2
fastnumbers>=5.1.0
fuzzywuzzy>=0.18.0
prompt-toolkit>=3.0.38
pyelftools>=0.31
python-levenshtein>=0.25.0
pyyaml>=5.3.1
pyzmq>=26.0.0
rapidyaml==0.5.0
sortedcontainers>=2.4.0
tabulate>=0.9.0
plotly.express>=4.14.3
pandas>=1.1.5
matplotlib>=3.3.4
